Implement two view-models
