+++ src/models/tabItem.js
@@ -1,5 +1,4 @@
 import { Record } from 'immutable';
-import yo from 'yo-yo';
 import Prefs from '../preference';
 import Constants from '../constants';
 
@@ -22,36 +21,6 @@ export default class TabItem extends TabItemRecord {
     this.faviconUrl = this.FAVICON_API + this.url;
   }
 
-  /**
-   * @function getDOM
-   * @return {object} yo-yo object
-   */
-  getDOM() {
-    const _openItem = (_e) => TabItem.openItem(this);
-    return yo`
-      <li class="ui-menu-item">
-        <div onclick=${_openItem} tabindex="-1" role="menuitem" class="ui-menu-item-wrapper">
-          <img src="${this.faviconUrl}" class="item-favicon" />
-          <div class="item-title">${this.title}</div>
-        </div>
-      </li>
-    `;
-  }
-
-  /**
-   * @function openItem
-   * @param {TabItem} tabItem a tabItem to open
-   */
-  static openItem(tabItem) {
-    if (Prefs.get(Constants.optionKeys.NO_NEW_TAB)) {
-      chrome.tabs.query({ active: true, currentWindow: true }, (tabs) => {
-        selectedTab = tabs[0];
-        chrome.tabs.update(selectedTab.id, { url: tabItem.url });
-      });
-    } else {
-      chrome.tabs.create({ url: tabItem.url, selected: false });
-    }
-  }
 
   /**
    * @function sort
