+++ src/popup.js
@@ -2,8 +2,7 @@ import I from 'immutable';
 import yo from 'yo-yo';
 import Constants from './constants';
 import Prefs from './preference';
-import TabItem from './models/tabItem';
-import ItemList from './models/itemList';
+import ItemList from './viewModels/itemList';
 import LocalRepository from './repositories/local';
 import SyncRepository from './repositories/sync';
 import 'jquery-ui-bundle/jquery-ui.min.js';
