+++ src/viewModels/itemList.js
@@ -1,9 +1,10 @@
 import yo from 'yo-yo';
 import Prefs from '../preference';
 import Constants from '../constants';
+import ListItem from './listItem';
 
 /**
- * Application Model for ItemList
+ * View-Model for ItemList
  */
 export default class ItemList {
   /**
@@ -26,7 +27,7 @@ export default class ItemList {
       <ul id="items-${this.target}" role="menu" tabindex="0"
         class="items ui-menu ui-widget ui-widget-content"
         style="height: ${popupHeight}px;">
-        ${this.items.map((item) => item.getDOM()).toArray()}
+        ${this.items.map((item) => ListItem.fromModel(item).getDOM()).toArray()}
       </ul>
     `;
   }
